Minor adjustments to My Museum Tour PDF cropping [WEB-2769]

+++ app/Console/Commands/GeneratePdfs.php
@@ -126,7 +126,9 @@ class GeneratePdfs extends Command
                 $fileName,
                 'public'
             );
-            unlink($localPath);
+            if (!\App::environment('local')) {
+                unlink($localPath);
+            }
         }
     }
